Use cache directory with array localisation cache by default
This can make short requests a bit faster and reduce DB pressure. It causes
localization cache to be cached via opcache instead of fetched
from DB.

Disable the code that makes $IP/cache a symbolic link and rm's .htaccess.
It is not reccomended that the cache directory be in the web directory.
It matters less for the "array" method of localisation cache, but for other
methods (which we are not using) this could result in private info being leaked.