Dockerfile: update versions of extensions
Generally, for skins and extensions that cloned the release-specific branch of
the repository, use the latest 1.41 commit. For skins and extensions that clone
a specific version of the repository, clone the latest version (which may not
be the latest commit). For extensions that previously cloned a version from
1.39 but that didn't have a 1.41 branch, clone based on a commit from the
master branch.

Additionally, update the list of extensions that are bundled with core, and
stop manually cloning `DiscussionTools`, `Echo`, `Linter`, and `Thanks`, which
are now bundled.