Update some extensions for 1.43 compatibility (#314)
* Update SemanticWatchlist to `cfd14a68f9cb073afc3f7a26f1c3baa071efbbd0`

We need to include SemanticMediaWiki/SemanticWatchlist#108 for 1.43
compatibility, I missed this during the 1.43 update of Taqasta. Upgrade to the
latest commit with CI passing.

* Update SubPageList to 3.0.1

We need to include ProfessionalWiki/SubPageList#77 for 1.43 compatibility, I
missed this during the 1.43 update of Taqasta.

* Update PubmedParser to v6.0.0

We need to include bovender/PubmedParser#16 for 1.43 compatibility.

* Update LinkTarget to latest `master` commit

We need to include wikimedia/mediawiki-extensions-LinkTarget@38af02fd38f15402031532c2d8bee784c61120df
to be able to modify the HTML output.

* Update BootstrapComponents to `476693d7a0f0de11bb67b775486950bff5cd1415`

We need to include patches like oetterer/BootstrapComponents#58 for 1.43
compatibility, use the latest version to include some other subsequent commits.

DEMO-137