Update SemanticMediaWiki and related extensions (#323)
* Update SemanticMediaWiki to 6.0.1
* Update SemanticCompoundQueries to 3.0.0
* Update SemanticResultFormats to 5.1.0
* Update SemanticScribunto to 2.3.3
* Update Maps to 12.0.0
* Update SemanticWatchlist to a newer commit
* Update SemanticDependencyUpdater to a newer commit

Note that Maps 12.x does not officially support SMW 6.x, but local testing
suggests that it should work fine, at least for basic SMW integration. If there
are issues, integration can be disabled by setting
`$egMapsDisableSmwIntegration = true;`.

SemanticDependencyUpdater is using a commit from the WikiTeq fork to add
support for triggering updates when a page is deleted; there is a PR pending to
add that support upstream.

WIK-2111