Remove GoogleDocCreator, MintyDocs, and Mpdf (#342)
GoogleDocCreator: we do not actually configure this anywhere, no need to
include extensions that we do not use.

MintyDocs: has a long-standing issue of calling methods on null, none of our
current clients use the extension.

Mpdf: extension pins to the `mpdf/mpdf` library with the version constraint
`^6.1`; the library used array and string offset access with curly braces
(deprecated in PHP 7.4, remove in PHP 8.0) until version 8.0.3, and also had
security issues until version 7.1.8. Avoid including a problematic extension
that we do not appear to use anywhere.

DEMO-137