Add and use support for adding patches via values.yml
Add a new top level key, `patches`, that holds an array of different patches to
apply, each with a `name` (the patch name in `_sources/patches`) and a `path`
(the path to apply the patch to, relative to `$MW_HOME`). For now, the patches
are only used for fixing extensions; support for applying patches to skins as
well can be added separately in the future if needed.

Migrate the two current extension patches, `FlexDiagrams.0.4.fix.diff` and
`GoogleLogin-fixes.patch`, to the new system.

WIK-2213